Refactor package structure to separate test runner from onit command codes.
