+++ Makefile
@@ -8,9 +8,13 @@ ONOS_TEST_DEBUG_VERSION := debug
 ONOS_BUILD_VERSION := stable
 
 build: # @HELP build the Go binaries and run all validations (default)
-build:
+build: build-onit
+
+build-onit:
 	go build -o build/_output/onit ./cmd/onit
-	go build -o build/_output/onit-k8s ./cmd/onit-k8s
+
+build-tests:
+	go build -o build/_output/onos-test-runner ./cmd/onos-test-runner
 
 test: # @HELP run the unit tests and source code validation
 test: build deps linters
@@ -34,19 +38,6 @@ deps: # @HELP ensure that the required dependencies are in place
 license_check: # @HELP examine and ensure license headers exist
 	./build/licensing/boilerplate.py -v
 
-onos-test-base-docker: # @HELP build onos-test base Docker image
-	@go mod vendor
-	docker build . -f build/base/Dockerfile \
-		--build-arg ONOS_BUILD_VERSION=${ONOS_BUILD_VERSION} \
-		-t onosproject/onos-test-base:${ONOS_TEST_VERSION}
-	@rm -rf vendor
-
-
-onos-test-docker: onos-test-base-docker # @HELP build onos-test Docker image
-	docker build . -f build/onos-test/Dockerfile \
-		--build-arg ONOS_CONTROL_BASE_VERSION=${ONOS_TEST_VERSION} \
-		-t onosproject/onos-tests:${ONOS_TEST_VERSION}
-
 # integration: @HELP build and run integration tests
 integration: kind
 	onit create cluster
@@ -54,21 +45,27 @@ integration: kind
 	onit add simulator
 	onit run suite integration-tests
 
+onos-test-runner-docker: # @HELP build onos-test-runner Docker image
+	@go mod vendor
+	docker build . -f build/onos-test-runner/Dockerfile \
+		--build-arg ONOS_BUILD_VERSION=${ONOS_BUILD_VERSION} \
+		-t onosproject/onos-test-runner:${ONOS_TEST_VERSION}
+	@rm -rf vendor
 
 images: # @HELP build all Docker images
-images: build onos-test-docker
+images: onos-test-runner-docker
 
 kind: # @HELP build Docker images and add them to the currently configured kind cluster
 kind: images
 	@if [ `kind get clusters` = '' ]; then echo "no kind cluster found" && exit 1; fi
-	kind load docker-image onosproject/onos-tests:${ONOS_TEST_VERSION}
+	kind load docker-image onosproject/onos-test-runner:${ONOS_TEST_VERSION}
 
 
 all: build images
 
 
 clean: # @HELP remove all the build artifacts
-	rm -rf ./build/_output ./vendor ./cmd/onit/onit ./cmd/onit-k8s/onit-k8s
+	rm -rf ./build/_output ./vendor
 
 help:
 	@grep -E '^.*: *# *@HELP' $(MAKEFILE_LIST) \
