+++ cmd/onos-test-runner/main.go
@@ -12,13 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package integration
+package main
 
 import (
+	"fmt"
 	"github.com/onosproject/onos-test/pkg/runner"
+	"github.com/onosproject/onos-test/test"
+	_ "github.com/onosproject/onos-test/test/atomix"
+	_ "github.com/onosproject/onos-test/test/integration"
+	"os"
 )
 
-var (
-	// Registry holds the registered test names
-	Registry = runner.NewRegistry()
-)
+func main() {
+	cmd := runner.GetOnosTestRunnerCommand(test.Registry)
+	if err := cmd.Execute(); err != nil {
+		fmt.Println(err)
+		os.Exit(1)
+	}
+}
