+++ go.mod
@@ -3,6 +3,7 @@ module github.com/onosproject/onos-test
 go 1.12
 
 require (
+	github.com/atomix/atomix-go-client v0.0.0-20190706023513-51518f9bf013
 	github.com/atomix/atomix-k8s-controller v0.0.0-20190620084759-d5e65f7fbf68
 	github.com/fatih/color v1.7.0
 	github.com/ghodss/yaml v1.0.0
