+++ pkg/onit/cli/completion.go
