+++ pkg/onit/console/spinner.go
