+++ pkg/onit/simulator.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package runner
+package onit
 
 import (
 	"encoding/json"
