+++ test/atomix/suites.go
@@ -12,22 +12,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package atomix
 
 import (
-	"fmt"
+	"github.com/onosproject/onos-test/pkg/runner"
 	"github.com/onosproject/onos-test/test"
-	_ "github.com/onosproject/onos-test/test/atomix"
-	_ "github.com/onosproject/onos-test/test/integration"
-	"os"
-
-	"github.com/onosproject/onos-test/pkg/onit/cli"
 )
 
-func main() {
-	cmd := cli.GetOnitCommand(test.Registry)
-	if err := cmd.Execute(); err != nil {
-		fmt.Println(err)
-		os.Exit(1)
-	}
+var (
+	// AtomixTests is the complete Atomix test suite
+	AtomixTests = runner.NewTestSuite("atomix")
+)
+
+func init() {
+	test.Registry.RegisterTestSuite(*AtomixTests)
 }
