+++ test/integration/singlepathtest.go
@@ -15,10 +15,11 @@
 package integration
 
 import (
+	"github.com/onosproject/onos-test/pkg/runner"
+	"github.com/onosproject/onos-test/test"
 	"testing"
 
-	"github.com/onosproject/onos-test/pkg/env"
-	"github.com/onosproject/onos-test/pkg/runner"
+	"github.com/onosproject/onos-test/test/env"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -69,5 +70,5 @@ func TestSinglePath(t *testing.T) {
 }
 
 func init() {
-	Registry.RegisterTest("single-path", TestSinglePath, []*runner.TestSuite{AllTests, IntegrationTests})
+	test.Registry.RegisterTest("single-path", TestSinglePath, []*runner.TestSuite{AllTests, IntegrationTests})
 }
