+++ test/registry.go
@@ -12,22 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package test
 
 import (
-	"fmt"
-	"github.com/onosproject/onos-test/test"
-	_ "github.com/onosproject/onos-test/test/atomix"
-	_ "github.com/onosproject/onos-test/test/integration"
-	"os"
-
-	"github.com/onosproject/onos-test/pkg/onit/cli"
+	"github.com/onosproject/onos-test/pkg/runner"
 )
 
-func main() {
-	cmd := cli.GetOnitCommand(test.Registry)
-	if err := cmd.Execute(); err != nil {
-		fmt.Println(err)
-		os.Exit(1)
-	}
-}
+var (
+	// Registry holds the registered test names
+	Registry = runner.NewRegistry()
+)
